requirements.txt
fastapi
deta
requests
feedparser
google-generativeai